+++ mlir/lib/Compiler/CompilerPipeline.cpp
@@ -147,6 +147,9 @@ QuantumCompilerPipeline::runPipeline(ModuleOp module,
   }
   // Stage 5: Optimization passes
   if (failed(runStage([&](PassManager& pm) {
+        if (config_.enableDecomposeMultiControlled) {
+          pm.addPass(qco::createDecomposeMultiControlled());
+        }
         if (!config_.disableMergeSingleQubitRotationGates) {
           pm.addPass(qco::createMergeSingleQubitRotationGates());
         }